Move userdic package under spellchecker
